% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.R
\name{cite}
\alias{cite}
\alias{cite_p}
\alias{cite_t}
\title{Cite reference(s) in a bibtex file}
\usage{
cite(
  key,
  type = c("t", "p"),
  bib,
  default_bib = Sys.getenv("BIB_NAME", unset = "bib")
)

cite_p(key, bib, default_bib = Sys.getenv("BIB_NAME", unset = "bib"))

cite_t(key, bib, default_bib = Sys.getenv("BIB_NAME", unset = "bib"))
}
\arguments{
\item{key}{Character vector; reference bibtex key(s)}

\item{type}{Character vector of length 1; type of citation. Must choose
"t" (textual citation) or "p" (parenthetical citation)}

\item{bib}{List of class 'bibentry'; bibliography read in from a bibtex file}

\item{default_bib}{Name of a bibliography object (list of class 'bibentry')
in the global environment to use instead of \code{bib}; defaults to "bib"}
}
\value{
Character vector of length 1; the formatted reference.
}
\description{
The bibliography should be read in from a \code{bibtex} file with
\code{\link[RefManageR:ReadBib]{RefManageR::ReadBib()}}
}
\details{
Either \code{bib} or \code{default_bib} should be provided, but not both
(see examples).

\code{cite_p()} is a shortcut for \code{cite(type = "p", ...)} and
\code{cite_t()} is a shortcut for \code{cite(type = "t", ...)}.
}
\examples{
# Load a bibliography
if (requireNamespace("bibtex", quietly = TRUE)) {
file.name <- system.file("Bib", "RJC.bib", package="RefManageR")
bib <- RefManageR::ReadBib(file.name)

# Check some of the keys (the name of each entry is its bibtex key)
head(names(bib))

# If an object called "bib" exists, it will be used by default for citing
cite(c("tekwe2013application", "sarkar2013adaptive"), "t")
# You can change the default name with the environmental variable "BIB_NAME"
my_bib <- bib
rm(bib)
Sys.setenv("BIB_NAME" = "my_bib")
cite("tekwe2013application", "t")
# Or provide the bibliography directly
cite("tekwe2013application", "t", bib = my_bib)

# Use cite_t() and cite_p() to avoid specifying "type"
cite_p("tekwe2013application")
cite_t("tekwe2013application")
} else {
  warning("Need bibtex package for example")
}
}
