% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_citations.R
\name{extract_citations}
\alias{extract_citations}
\title{Extract citations (formatted like \verb{[@key]}) from an Rmd file}
\usage{
extract_citations(rmd_file)
}
\arguments{
\item{rmd_file}{Character vector; path to Rmd file}
}
\value{
Dataframe (tibble) with one column 'key'
}
\description{
Extract citations (formatted like \verb{[@key]}) from an Rmd file
}
\examples{
tempfile <- tempfile(fileext = ".Rmd")
lines <- c(
  "---",
  "title: 'Report'",
  "output_format: html_document",
  "---",
  "This statement is supported by @Nitta2021",
  "[@Foo1983; @Blah2001; but see @Smith2020]."
)
write(lines, tempfile)
extract_citations(tempfile)
file.remove(tempfile)
}
