% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_refs_bib.R
\name{filter_refs_bib}
\alias{filter_refs_bib}
\title{Filter a list of references in a bibtex file to those occurring in an
Rmd file}
\usage{
filter_refs_bib(
  rmd_file,
  bib_in = "main_library.bib",
  bib_out = "references.bib",
  silent = FALSE
)
}
\arguments{
\item{rmd_file}{Character vector; Path to Rmd file(s)}

\item{bib_in}{String or list; if string, the path to the bibtex file to
filter. If list, should be result of reading in a bibtex file with
\code{\link[RefManageR:ReadBib]{RefManageR::ReadBib()}}}

\item{bib_out}{Path to write filtered bibtex file}

\item{silent}{Logical; should a warning be issued for missing references?}
}
\value{
NULL; externally, the filtered bibtex will be written to \code{bib_out}
}
\description{
Filter a list of references in a bibtex file to those occurring in an
Rmd file
}
\examples{
if (requireNamespace("bibtex", quietly = TRUE)) {

  # Write out an example Rmd file
  rmd_tempfile <- tempfile(fileext = ".Rmd")
  lines <- c(
    "---",
    "title: 'Report'",
    "output_format: html_document",
    "---",
    "This statement is supported by @tekwe2013application",
    "[@sarkar2013adaptive]."
  )
  write(lines, rmd_tempfile)

  # Filter the bibliography
  bib_tempfile <- tempfile(fileext = ".bib")
  filter_refs_bib(
    rmd_file = rmd_tempfile,
    bib_in = system.file("Bib", "RJC.bib", package = "RefManageR"),
    bib_out = bib_tempfile
  )
  # Check that it worked
  readLines(bib_tempfile)

  # Cleanup
  file.remove(rmd_tempfile)
  file.remove(bib_tempfile)

} else {
  warning("Need bibtex package for example")
}
}
