% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_refs_yaml.R
\name{filter_refs_yaml}
\alias{filter_refs_yaml}
\title{Filter a list of references in YAML format to those occurring in an Rmd file}
\usage{
filter_refs_yaml(
  rmd_file,
  yaml_in = "ms/main_library.yaml",
  yaml_out = "ms/references.yaml",
  silent = FALSE
)
}
\arguments{
\item{rmd_file}{Character vector; Path to Rmd file(s)}

\item{yaml_in}{String or list; if string, the path to the YAML file to
filter. If list, should be result of reading in a YAML file with
yaml::read_yaml()}

\item{yaml_out}{Path to write filtered YAML reference file}

\item{silent}{Logical; should a warning be issued for missing references?}
}
\value{
NULL; externally, the filtered YAML will be written to \code{yaml_out}
}
\description{
Filter a list of references in YAML format to those occurring in an Rmd file
}
