% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_pd.R
\name{cite_pd}
\alias{cite_pd}
\title{Cite a reference using pandoc}
\usage{
cite_pd(ref, bib, csl, context = NULL, format = "plain", glue = TRUE)
}
\arguments{
\item{ref}{Character vector of length 1; a character string including
the reference to cite, in pandoc format. See examples.}

\item{bib}{Character vector; path to bibliography file(s) in \code{bib} or \code{yaml}
format.}

\item{csl}{Character vector of length 1; path to CSL file.}

\item{context}{Character vector; path to file(s) with citations in pandoc
format. These aren't cited, but used
for disambiguating references (see Details).}

\item{format}{Character vector of length 1; pandoc output format. Must be
a valid output format for the pandoc \code{-t} argument. For a complete list,
see https://pandoc.org/MANUAL.html#general-options}

\item{glue}{Logical vector of length 1; should the output be passed through
\code{\link[glue:glue]{glue::glue()}}? Generally results in nicer printing (e.g., line breaks
show up on different lines instead of \verb{\\n}). Default \code{TRUE}.}
}
\value{
Character vector.
}
\description{
Unlike the normal use of pandoc to render \code{Rmd} to an output format, this
just returns a single formatted reference.
}
\details{
This function can be used to generate pandoc-formatted references in formats
that normally don't support pandoc, such as slides with xaringan or
CSV files.

The pandoc format for citing references is \verb{@<key>}, where \verb{<key>} is the
reference key (e.g., \verb{@Nitta2011} or \verb{[@Nitta2011]}).

Note that since pandoc isn't seeing a full (R)md file, it won't
automatically disambiguate references (cases where an 'a', 'b', etc. needs
to be added to differentiate between publications whose author and
year are the same). If you need to distinguish between such publications
that are cited in the same file (or set of files), provide the file path(s)
to \code{context}.
}
\examples{
# Specify example bibliography file (yaml format)
my_bib <- system.file(
  "extdata", "refs.yaml", package = "rmdref", mustWork = TRUE)

# Download a CSL file to a temporary location
temp_csl <- tempfile(fileext = ".csl")
utils::download.file(
  "https://raw.githubusercontent.com/citation-style-language/styles/master/ecology.csl", # nolint
  temp_csl
)

# Cite the reference (note no disambiguation)
cite_pd("@Nitta2011a", my_bib, temp_csl)

# Provide context for disambiguation: a CSV file with some cited references
csv_file <- system.file(
  "extdata", "data.csv", package = "rmdref", mustWork = TRUE)
read.csv(csv_file)

cite_pd("@Nitta2011a", my_bib, temp_csl, csv_file)

# Delete the temporary file
unlink(temp_csl)

}
