% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_pd.R
\name{cite_pd}
\alias{cite_pd}
\alias{bib_pd}
\title{Cite a reference using pandoc}
\usage{
cite_pd(
  ref,
  bib,
  csl,
  context = NULL,
  format = "plain",
  glue = TRUE,
  wrap = "auto",
  columns = 72,
  other_pd_args = NULL,
  default_bib = Sys.getenv("BIB_FILE", unset = "bib"),
  default_csl = Sys.getenv("CSL_FILE", unset = "csl"),
  default_context = Sys.getenv("CONTEXT_FILE", unset = "context")
)

bib_pd(
  ref = NULL,
  bib,
  csl,
  context = NULL,
  format = "plain",
  glue = TRUE,
  wrap = "auto",
  columns = 72,
  other_pd_args = NULL,
  default_bib = Sys.getenv("BIB_FILE", unset = "bib"),
  default_csl = Sys.getenv("CSL_FILE", unset = "csl"),
  default_context = Sys.getenv("CONTEXT_FILE", unset = "context")
)
}
\arguments{
\item{ref}{Character vector of length 1; a character string including
one or more citations in pandoc format used to create a citation
(\code{cite_pd()}) or bibliography entry (\code{bib_pd()}).}

\item{bib}{Character vector; path to bibliography file(s) in \code{bib} or \code{yaml}
format.}

\item{csl}{Character vector of length 1; path to CSL file.}

\item{context}{Character vector; path to file(s) with citations in pandoc
format. If used with \code{cite_pd()}, these aren't cited, but used
for disambiguating references (see Details). If used with \code{bib_pd()}, all
references in the file(s) will appear in the bibliography.}

\item{format}{Character vector of length 1; pandoc output format. Must be
a valid output format for the pandoc \code{-t} argument. For a complete list,
see https://pandoc.org/MANUAL.html#general-options}

\item{glue}{Logical vector of length 1; should the output be passed through
\code{\link[glue:glue]{glue::glue()}}? Generally results in nicer printing (e.g., line breaks
show up on different lines instead of \verb{\\n}). Default \code{TRUE}.}

\item{wrap}{Character vector of length 1; type of line wrapping in output.
With \code{auto} (the default), pandoc will attempt to wrap lines to the column
width specified by \code{columns} (default 72). With \code{none}, pandoc will not
wrap lines at all. With \code{preserve}, pandoc will attempt to preserve the
wrapping from the source document.}

\item{columns}{Numeric vector of length 1; width of lines when wrapping.}

\item{other_pd_args}{Character vector; other arguments to pass to pandoc.
Must be formatted with one element per word
(as in \code{args} for \code{\link[processx:run]{processx::run()}}).}

\item{default_bib}{Name of a character vector object in the global
environment to use for \code{bib}; defaults to \code{"bib"}. Can be changed by
setting environmental variable \code{BIB_FILE}.}

\item{default_csl}{Name of a character vector object in the global
environment to use for \code{csl}; defaults to \code{"csl"}. Can be changed by
setting environmental variable \code{CSL_FILE}.}

\item{default_context}{Name of a character vector object in the global
environment to use for \code{context}; defaults to \code{"context"}. Can be changed by
setting environmental variable \code{CONTEXT_FILE}.}
}
\value{
Character vector.
}
\description{
Unlike the normal use of pandoc to render \code{Rmd} to an output format, this
just returns a formatted citation or bibliography.
}
\details{
\code{cite_pd()} creates citations; \code{bib_pd()} creates a bibliography
(list of references).

This function can be used to generate pandoc-formatted references in formats
that normally don't support pandoc, such as slides with xaringan or
CSV files.

The pandoc format for citing references is \verb{@<key>}, where \verb{<key>} is the
reference key (e.g., \verb{@Nitta2011} or \verb{[@Nitta2011]}).

Note that since pandoc isn't seeing a full (R)md file, it won't automatically
disambiguate references when generating a citation (cases where an 'a', 'b',
etc. needs to be added to differentiate between publications whose author and
year are the same). If you need to distinguish between such publications that
are cited in the same file (or set of files), provide the file path(s) to
\code{context}.
}
\examples{
# Specify example bibliography file (yaml format)
bib <- system.file(
  "extdata", "refs.yaml", package = "rmdref", mustWork = TRUE)

# Download a CSL file to a temporary location
csl <- tempfile(fileext = ".csl")
utils::download.file(
  "https://raw.githubusercontent.com/citation-style-language/styles/master/ecology.csl", # nolint
  csl
)

# Cite the reference (note no disambiguation)
cite_pd("@Nitta2011a", bib = bib, csl = csl)

# Since the `bib` and `csl` objects match their default names for `cite_pd()`
# they can be omitted from the function call.
cite_pd("@Nitta2011a")

# Provide context for disambiguation: a CSV file with some cited references
csv_file <- system.file(
  "extdata", "data.csv", package = "rmdref", mustWork = TRUE)
read.csv(csv_file)

cite_pd("@Nitta2011a", context = csv_file)

# Make a list of references in the CSV file
bib_pd(context = csv_file, wrap = "none")

# Delete the temporary file
unlink(csl)

}
